Refactor scanner
