+++ lib/scanner/common/CMakeLists.txt
@@ -1,0 +1,9 @@
+add_library(cloxscanner_common INTERFACE)
+
+target_include_directories(
+  cloxscanner_common
+  INTERFACE
+    include
+)
+
+add_subdirectory(test)
