--- lib/test/CMakeLists.txt
@@ -1,2 +1,0 @@
-include(Utils)
-IncludeSubdirectories(${BUILD_TESTING})
